# lib/api-zod/src/generated/types/deleteResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shadow Leveling - Gamified Habit Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface DeleteResult {
  success: boolean;
}